% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparePoints.R
\name{cp_style}
\alias{cp_style}
\title{Style a comparePoints visualization}
\usage{
cp_style(comparePoints, axis_format = ",", jitter_width = 0)
}
\arguments{
\item{axis_format}{option to pass to d3.format()}

\item{jitter_width}{jitter width in pixels}
}
\description{
Style a comparePoints visualization
}
