% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_quantiles.R
\name{draw_quantiles}
\alias{draw_quantiles}
\title{Draw quantile lines}
\usage{
draw_quantiles(pinpoint, probs = c(0.25, 0.75), color = "grey")
}
\arguments{
\item{pinpoint}{a pinpoint visualization}

\item{probs}{numeric vector of probabilities with values between 0 and 1}

\item{color}{color of quantile lines}
}
\description{
Draw quantile lines
}
