% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint.R
\name{pinpoint}
\alias{pinpoint}
\title{Create a pinpoint visualization}
\usage{
pinpoint(data, x, fill = NULL, tooltip = NULL,
  compare_to = "diff_from_mean", title = "")
}
\arguments{
\item{data}{A table of data}

\item{x}{A numeric variable.}

\item{fill}{Point color. Either an unquoted column name or valid string/hex color}

\item{tooltip}{Variable to display within tooltip}

\item{compare_to}{the x-intercept to compare against values. Either a numeric value or "diff_from_mean", "diff_from_median", or "z-score"}

\item{title}{Visualization title}
}
\description{
Create a pinpoint visualization
}
