% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_create_app}
\alias{pinpoint_create_app}
\title{Creates or updates an app}
\usage{
pinpoint_create_app(CreateApplicationRequest)
}
\arguments{
\item{CreateApplicationRequest}{[required]}
}
\description{
Creates or updates an app.
}
\section{Request syntax}{
\preformatted{pinpoint$create_app(
  CreateApplicationRequest = list(
    Name = "string",
    tags = list(
      "string"
    )
  )
)
}
}

\keyword{internal}
