% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_delete_apns_sandbox_channel}
\alias{pinpoint_delete_apns_sandbox_channel}
\title{Delete an APNS sandbox channel}
\usage{
pinpoint_delete_apns_sandbox_channel(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Delete an APNS sandbox channel.
}
\section{Request syntax}{
\preformatted{pinpoint$delete_apns_sandbox_channel(
  ApplicationId = "string"
)
}
}

\keyword{internal}
