% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_delete_gcm_channel}
\alias{pinpoint_delete_gcm_channel}
\title{Deletes the GCM channel for an app}
\usage{
pinpoint_delete_gcm_channel(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Deletes the GCM channel for an app.
}
\section{Request syntax}{
\preformatted{pinpoint$delete_gcm_channel(
  ApplicationId = "string"
)
}
}

\keyword{internal}
