% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_delete_user_endpoints}
\alias{pinpoint_delete_user_endpoints}
\title{Deletes endpoints that are associated with a User ID}
\usage{
pinpoint_delete_user_endpoints(ApplicationId, UserId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{UserId}{[required] The unique ID of the user.}
}
\description{
Deletes endpoints that are associated with a User ID.
}
\section{Request syntax}{
\preformatted{pinpoint$delete_user_endpoints(
  ApplicationId = "string",
  UserId = "string"
)
}
}

\keyword{internal}
