% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_voice_channel}
\alias{pinpoint_get_voice_channel}
\title{Get a Voice Channel}
\usage{
pinpoint_get_voice_channel(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Get a Voice Channel
}
\section{Request syntax}{
\preformatted{pinpoint$get_voice_channel(
  ApplicationId = "string"
)
}
}

\keyword{internal}
