% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_update_apns_voip_channel}
\alias{pinpoint_update_apns_voip_channel}
\title{Update an APNS VoIP channel}
\usage{
pinpoint_update_apns_voip_channel(APNSVoipChannelRequest, ApplicationId)
}
\arguments{
\item{APNSVoipChannelRequest}{[required]}

\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Update an APNS VoIP channel
}
\section{Request syntax}{
\preformatted{pinpoint$update_apns_voip_channel(
  APNSVoipChannelRequest = list(
    BundleId = "string",
    Certificate = "string",
    DefaultAuthenticationMethod = "string",
    Enabled = TRUE|FALSE,
    PrivateKey = "string",
    TeamId = "string",
    TokenKey = "string",
    TokenKeyId = "string"
  ),
  ApplicationId = "string"
)
}
}

\keyword{internal}
