% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_update_voice_channel}
\alias{pinpoint_update_voice_channel}
\title{Update an Voice channel}
\usage{
pinpoint_update_voice_channel(ApplicationId, VoiceChannelRequest)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{VoiceChannelRequest}{[required]}
}
\description{
Update an Voice channel
}
\section{Request syntax}{
\preformatted{pinpoint$update_voice_channel(
  ApplicationId = "string",
  VoiceChannelRequest = list(
    Enabled = TRUE|FALSE
  )
)
}
}

\keyword{internal}
