% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.pp_df}
\alias{plot.pp_df}
\title{Plot processed Swift fixes from PinPoint Host software output}
\usage{
\method{plot}{pp_df}(pp_df, min_sats = 4, max_hdop = 20)
}
\arguments{
\item{pp_df}{a `pp_df` object (i.e., a `data.frame` created by \code{\link{read_pp_swift}})}

\item{min_sats}{numeric/integer scalar indicating the minimum number of satellites
required (default is 4) to accept a GPS fix as valid.}

\item{max_hdop}{numeric scalar indicating the maximum dilution of precision (default is
<= 20) to accept a GPS fix as valid.}
}
\description{
This function uses \code{\link[leaflet]{leaflet}} to generate a basic interactive map of
 PinPoint Swift fixes from the `pp_df` class data produced by \code{\link{read_pp_swift}}.
 A separate overlay group is created for each distinct tag ID.  Failed GPS
 fix attempts are removed automatically.  Users can optionally filter GPS fixes by the
 reported dilution of precision `max_hdop` or the number of satellites `min_sats`.
 Limited experimentation finds fixes with at least 4 satellites and an HDOP <= 20 and to
 be of adequate locational quality, with an RMS of about 20 m (see details of
 \code{\link{test_pp_swift}} for more information).  These are the defaults.  For larger
 scale studies, fixes from 3 satellites may also be useful.
}
\examples{
\dontrun{
# Select appropriate text files from file selection window that opens
pp_tests <- list.files(path = system.file("extdata", package = "pinpoint"), full.names = TRUE)
dat <- read_pp_swift(pp_tests)
plot(dat)
plot(dat, min_sats = 8)
}
}
