% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pp_swift.R
\name{plot_pp_swift}
\alias{plot_pp_swift}
\title{Plot processed Swift fixes from PinPoint Host software output}
\usage{
plot_pp_swift(pp_swift, out_tz = "America/New_York", max_hdop = 4,
  min_sats = 4)
}
\arguments{
\item{pp_swift}{either (1) a character path to the *.txt file created by the PinPoint
Host software when processing Swift fixes or (2) a data frame created by
\code{\link{read_pp_swift}}}

\item{out_tz}{character string indicating the desired output \code{\link[base]{timezone}}
of the GPS fixes.  Datetimes will be converted from the standard GMT time zone of
PinPoint fixes. Default is America/New_York.}

\item{max_hdop}{numeric scalar indicating the maximum dilution of precision (default is
<= 4) to accept a GPS fix as valid.}

\item{min_sats}{numeric/integer scalar indicating the minimum number of satellites
required (default is 4) to accept a GPS fix as valid.}
}
\description{
This function plots fixes from either (1) the text file generated by the
 PinPoint Host software or (2) the data frame generated by \code{\link{read_pp_swift}}
 and generates a basic interactive map of fixes using \code{\link[leaflet]{leaflet}}
 with a separate overlay group for each distinct tag ID (in the case a combined data
 frame of \code{\link{read_pp_swift}} output is passed to the function).  Failed GPS
 fix attempts are removed automatically.  Users can optionally filter GPS fixes by the
 reported dilution of precision `max_hdop` or the number of satellites `min_sats`.
 Limited experimentation finds fixes with HDOP <= 4 and at least 4 satellites to be of
 adequate locational quality (usually within 20 m).  These are the defaults.
}

