% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint.R
\name{pp_style}
\alias{pp_style}
\title{Style a pinpoint visualization}
\usage{
pp_style(pinpoint, point_radius = 10, number_format = ".5",
  jitter_width = 0, fill_colors = NULL, compare_mark_color = "black",
  greater_than_color = "forestgreen", less_than_color = "firebrick",
  diff_line_type = "dashed", axis_range = NULL, ticks = 8)
}
\arguments{
\item{pinpoint}{A pinpoint object}

\item{point_radius}{Radius of points (in pixels)}

\item{number_format}{option to pass to d3.format()}

\item{jitter_width}{jitter width in pixels}

\item{fill_colors}{fill colors}

\item{compare_mark_color}{color of compare mark}

\item{greater_than_color}{color of diff line and text when value is greater than compare_mark_color}

\item{less_than_color}{color of diff line and text when value is less than compare_mark_color}

\item{diff_line_type}{'solid' or 'dashed'}

\item{axis_range}{vector of length two, min and maximium range of axis}

\item{ticks}{number of axis ticks}
}
\description{
Style a pinpoint visualization
}
