% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pp_swift.R
\name{read_pp_swift}
\alias{read_pp_swift}
\title{Read processed Swift fixes from PinPoint Host software output}
\usage{
read_pp_swift(swift_txt = NULL, out_tz = "America/New_York",
  valid_only = FALSE)
}
\arguments{
\item{swift_txt}{character scalar or vector containing path(s) to the *.txt file created
by the PinPoint Host software when processing Swift fixes.  Default (`NULL`) prompts the
user to select one or more files.}

\item{out_tz}{character string indicating the desired output \code{\link[base]{timezone}}
of the GPS fixes.  Datetimes will be converted from the standard GMT time zone of
PinPoint fixes. Default is America/New_York.}

\item{valid_only}{logical indicating whether unsuccessful GPS fix attempts should be
removed from the output data frame.  The default (FALSE) retains the scheduled fixes
that were unsuccessful.}
}
\description{
This function reads the text file generated by the PinPoint Host software and
 generates a tidy data frame for further analysis.  The user can optionally
 remove unsuccessful fixes.
}

