% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_pp_swift.R
\name{test_pp_swift}
\alias{test_pp_swift}
\title{Test locational accuracy of PinPoint Swift fixes against a reference coordinate}
\usage{
test_pp_swift(pp_df, ref_coords = c(-83.36, 33.95), min_sats = 4,
  max_hdop = 20)
}
\arguments{
\item{pp_df}{a `pp_df` object (i.e., a `data.frame` created by \code{\link{read_pp_swift}})}

\item{ref_coords}{a vector of coordinates (decimal degrees, WGS84 datum) of the reference location
in the form (longitude, latitude).}

\item{min_sats}{numeric/integer scalar indicating the minimum number of satellites
required (default is 4) to accept a GPS fix as valid.}

\item{max_hdop}{numeric scalar indicating the maximum dilution of precision (default is
<= 20) to accept a GPS fix as valid.}
}
\value{
A histogram of locational errors (with median highlighted) and a report of 2-D RMS
 error by tag ID, and a `pp_df` object containing the input data and two new columns:
\itemize{
  \item error_l - the absolute locational error, in meters
  \item error_b - the bearing of the locational error, in degrees from north
}
}
\description{
This function evaluates the locational accuracy of PinPoint Swift fixes by comparing
 them to an input reference coordinate (i.e., tag location in lat/long decimal degrees
 WGS84).  The most convenient input is a `pp_df` class data.frame produced by
 \code{\link{read_pp_swift}}. Failed GPS fix attempts are removed automatically.  Users
 can optionally filter GPS fixes by the reported horizontal dilution of precision `max_hdop`
 or the number of satellites `min_sats`.  Limited experimentation finds fixes with at least
 4 satellites and an HDOP <= 20 and to be of adequate locational quality, with an RMS of
 about 20 m (see details).  These are the defaults.  For larger scale studies, fixes from
 3 satellites may also be useful.
}
\details{
The reported 2-D RMS error assumes that errors in the north-south and east-west directions
 are equivalent (i.e., that locational fixes describe a circular distribution around the
 true position; likely reasonable with >= 4 satellites).  Assuming also that mean positional
 error is zero (or close to it), RMS is essentially equivalent to 1 standard devation, which
 in the two-directional case means that a circle of radius RMS is expected to contain about
 63\% of positional locations on average; double it for the 95\% equivalent.  See Diggelen
 (2007) for more details: \url{http://gpsworld.com/gpsgnss-accuracy-lies-damn-lies-and-statistics-1134/})
}

