% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparePoints.R
\name{comparePoints}
\alias{comparePoints}
\title{Create a comparePoints visualization}
\usage{
comparePoints(data, x, fill = NULL, compare_mark = "diff_from_mean",
  title = "", ...)
}
\arguments{
\item{data}{A table of data}

\item{x}{A numeric variable.}

\item{fill}{dot color. Either an unquoted column name or valid string or hex color}

\item{compare_mark}{the x-intercept to compare against values. Either a numeric value or "diff_from_mean", "diff_from_median", or "z-score"}

\item{title}{Visualization title}

\item{...}{optional variables to include in the tooltip}
}
\description{
Create a comparePoints visualization
}
