% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparePoints.R
\name{cp_style}
\alias{cp_style}
\title{Style a comparePoints visualization}
\usage{
cp_style(comparePoints, number_format = ".5", jitter_width = 0,
  fill_colors = NULL, compare_mark_color = "black",
  greater_than_color = "forestgreen", less_than_color = "firebrick",
  diff_line_type = "solid")
}
\arguments{
\item{number_format}{option to pass to d3.format()}

\item{jitter_width}{jitter width in pixels}

\item{fill_colors}{fill colors}

\item{compare_mark_color}{color of compare mark}

\item{greater_than_color}{color of diff line and text when value is greater than compare_mark_color}

\item{less_than_color}{color of diff line and text when value is less than compare_mark_color}

\item{diff_line_type}{'solid' or 'dashed'}
}
\description{
Style a comparePoints visualization
}
