% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_deviations.R
\name{draw_deviations}
\alias{draw_deviations}
\title{Draw standard deviation lines from the mean}
\usage{
draw_deviations(pinpoint, deviations = 1, color = "red")
}
\arguments{
\item{pinpoint}{a pinpoint visualization}

\item{deviations}{either 1, 2, or 3}

\item{color}{color of deviation lines}
}
\description{
Draw standard deviation lines from the mean
}
